requirements.txt
beautifulsoup4==4.11.2
pymongo==4.3.2
python-dotenv==1.0.0
requests==2.25.1
tqdm==4.64.1
